# Conflicts:
#	SPServer/src/main/resources/contract2.txt

+++ SPServer/src/main/java/com/himedia/spserver/repository/ShopCategoryRepository.java
@@ -3,5 +3,5 @@ package com.himedia.spserver.repository;
 import com.himedia.spserver.entity.SHOP.SHOP_Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShopCategoryRepository extends JpaRepository<SHOP_Category,Integer> {
+public interface ShopCategoryRepository extends JpaRepository<SHOP_Category,Long> {
 }
